chore: refactored Notifications out for maintainability
