+++ frontend/src/components/useNotifications/NotificationsProvider.tsx
@@ -1,105 +1,18 @@
-import { useState, useContext, useMemo, useCallback } from "react";
+import { useState, useMemo, useCallback } from "react";
 import type { ReactNode } from "react";
-import Snackbar from "@mui/material/Snackbar";
-import SnackbarContent from "@mui/material/SnackbarContent";
-import IconButton from "@mui/material/IconButton";
-import CloseIcon from "@mui/icons-material/Close";
-import Alert from "@mui/material/Alert";
-import Badge from "@mui/material/Badge";
+import Notification from "./Notification";
 import { NotificationsContext } from "./NotificationsContext";
+import type {
+  ShowNotificationOptions,
+  ShowNotification,
+  CloseNotification,
+  RemoveNotification,
+} from "./useNotifications";
 
 const generateId = (() => {
   let nextId = 0;
   return () => nextId++;
 })();
-};
-
-export interface ShowNotificationOptions {
-  key?: string;
-  severity?: "info" | "warning" | "error" | "success";
-  autoHideDuration?: number;
-  actionText?: React.ReactNode;
-  onAction?: () => void;
-}
-
-export interface NotificationQueueItem {
-  notificationKey: string;
-  message: string;
-  isOpen: boolean;
-  options: ShowNotificationOptions;
-}
-
-interface NotificationProps extends NotificationQueueItem {
-  badge?: string;
-}
-
-const Notification = ({
-  notificationKey,
-  message,
-  isOpen,
-  options,
-  badge,
-}: NotificationProps) => {
-  const { close, remove } = useContext(NotificationsContext)!;
-  const handleClose = useCallback(
-    (_event: unknown, reason?: string) => {
-      if (reason === "clickaway") {
-        return;
-      }
-      close(notificationKey);
-    },
-    [notificationKey, close],
-  );
-
-  const handleExited = useCallback(() => {
-    remove(notificationKey);
-  }, [notificationKey, remove]);
-
-  const action = (
-    <>
-      <IconButton
-        size="small"
-        aria-label="Close"
-        title="Close"
-        color="inherit"
-        onClick={handleClose}
-      >
-        <CloseIcon fontSize="small" />
-      </IconButton>
-    </>
-  );
-
-  const slotProps = {
-    transition: {
-      onExited: handleExited,
-    },
-  };
-
-  return (
-    <Snackbar
-      key={notificationKey}
-      anchorOrigin={{ vertical: "top", horizontal: "center" }}
-      open={isOpen}
-      autoHideDuration={options.autoHideDuration}
-      onClose={handleClose}
-      slotProps={slotProps}
-    >
-      <Badge badgeContent={badge} color="primary" sx={{ width: "100%" }}>
-        {options.severity ? (
-          <Alert
-            severity={options.severity}
-            sx={{ width: "100%" }}
-            action={action}
-          >
-            {message}
-          </Alert>
-        ) : (
-          <SnackbarContent message={message} action={action} />
-        )}
-      </Badge>
-    </Snackbar>
-  );
-};
 
 interface NotificationsProps {
   queue: Array<NotificationQueueItem>;
@@ -114,14 +27,12 @@ const Notifications = ({ queue }: NotificationsProps) => {
   ) : null;
 };
 
-export type ShowNotification = (
-  message: string,
-  options: ShowNotificationOptions,
-) => string;
-
-export type CloseNotification = (id: string) => void;
-
-export type RemoveNotification = (id: string) => void;
+export interface NotificationQueueItem {
+  notificationKey: string;
+  message: string;
+  isOpen: boolean;
+  options: ShowNotificationOptions;
+}
 
 export interface NotificationsProviderProps {
   children?: ReactNode;
